15_udemy_agentic_ai/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
backoff==2.2.1
baseblock==0.2.24
cachetools==6.2.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
croniter==6.0.0
cryptography==38.0.1
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.8.0
fastapi==0.118.2
frozenlist==1.8.0
fsspec==2024.12.0
google-auth==2.41.1
google-genai==1.41.0
greenlet==3.2.4
grpcio==1.75.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.2
hyperframe==6.1.0
idna==3.10
jiter==0.11.0
json_repair==0.44.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.79
langchain-experimental==0.3.4
langchain-mongodb==0.7.0
langchain-neo4j==0.5.0
langchain-openai==0.3.35
langchain-qdrant==0.2.1
langchain-text-splitters==0.3.11
langgraph==0.6.10
langgraph-checkpoint==2.1.2
langgraph-checkpoint-mongodb==0.2.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.34
lark==1.3.0
marshmallow==3.26.1
mem0ai==0.1.118
multidict==6.7.0
mypy_extensions==1.1.0
neo4j==5.28.2
neo4j-graphrag==1.10.0
numpy==2.3.3
openai==0.27.10
openai-helper==0.2.4
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
portalocker==3.2.0
posthog==6.7.6
propcache==0.4.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
pymongo==4.15.3
pypdf==6.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==311
PyYAML==6.0
qdrant-client==1.15.1
rank-bm25==0.2.2
redis==6.4.0
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rq==2.6.0
rsa==4.9.1
scipy==1.16.2
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.14.3
SQLAlchemy==2.0.43
starlette==0.48.0
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
types-PyYAML==6.0.12.20250915
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
unicodedata2==16.0.0
urllib3==2.5.0
uvicorn==0.37.0
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
